Context and graph are gone, next step: handle the LazyTensor in python
