+++ ttnn/CMakeLists.txt
@@ -76,11 +76,11 @@ target_sources(
         core/graph/graph_argument_serializer.cpp
         core/graph/graph_processor.cpp
         core/graph/graph_trace_utils.cpp
-        core/experimental/jit/node.cpp
-        core/experimental/jit/context.cpp
         core/old_infra_device_operation.cpp
         core/reports.cpp
         core/run_operation.cpp
+        core/tensor/experimental/jit/lazy_tensor.cpp
+        core/tensor/experimental/jit/graph_utils.cpp
         core/tensor/flatbuffer/tensor_flatbuffer.cpp
         core/tensor/flatbuffer/tensor_spec_flatbuffer.cpp
         core/tensor/host_buffer/functions.cpp
@@ -100,7 +100,6 @@ target_sources(
         core/tensor/types.cpp
         core/tensor/unit_mesh/unit_mesh_utils.cpp
         core/tensor/xtensor/partition.cpp
-        core/tensor/lazy_tensor.cpp
 )
 target_include_directories(
     ttnn_core
